+++ mediawiki/__init__.py
@@ -4,7 +4,8 @@ mediawiki module initialization
 from .mediawiki import (MediaWiki, MediaWikiPage, URL, VERSION)
 from .exceptions import (MediaWikiException, PageError, MediaWikiGeoCoordError,
                          RedirectError, DisambiguationError,
-                         MediaWikiAPIURLError, HTTPTimeoutError)
+                         MediaWikiAPIURLError, HTTPTimeoutError,
+                         MediaWikiCategoryTreeError)
 
 __author__ = 'Tyler Barrus'
 __maintainer__ = 'Tyler Barrus'
@@ -16,4 +17,5 @@ __url__ = URL
 
 __all__ = ['MediaWiki', 'PageError', 'RedirectError', 'MediaWikiException',
            'DisambiguationError', 'MediaWikiAPIURLError',
-           'HTTPTimeoutError', 'MediaWikiGeoCoordError']
+           'HTTPTimeoutError', 'MediaWikiGeoCoordError',
+           'MediaWikiCategoryTreeError']