+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Current
 
+### Version 0.3.10
+
+* Add categorytree support
+* Remove adding 'http:' to references if missing
+
 ### Version 0.3.9
 
 * Fix infinite loop on continued queries: [issue #15](https://github.com/barrust/mediawiki/issues/15)